Move test to avoid dependencies
